requirements.txt
asgiref==3.4.1
certifi==2021.5.30
charset-normalizer==2.0.4
Django==3.2.7
django-ckeditor==6.1.0
django-js-asset==1.2.2
djangorestframework==3.12.4
idna==3.2
mysqlclient==2.0.3
pytz==2021.1
requests==2.26.0
sqlparse==0.4.1
urllib3==1.26.6
